Injects a content script to act as host on the page.
A lot of this comes from the same article about running an Angular app in a Chrome extension and how to work with builders: https://medium.com/angular-in-depth/chrome-extension-with-angular-why-and-how-778200b87575

I realized that the background script was loaded but not actually executed because I was not including the Webpack runtime.

I also included types for Chrome extensions so they would play nicely with TypeScript.