Adds an iframe to an extension page to render a sticky note.
This required the URL to be marked "web accessible" in the manifest. I implemented this with basic Chrome extension APIs and without tests. I'll switch this over to WebExtensions and shard it to be more testable.

I also made minor edits to `tslint` and `tsconfig`. I'm not sure why it isn't `strict` by default.